+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Fri Feb  9 11:00:06 UTC 2018
+
+* Add config_parsing_failed diagnostic event
+
 # Thu Feb  8 12:36:42 UTC 2018
 
 * Lunar new year theme (groups  5 & 6)